Added timer:
New Timing mechanism that uses system.millis time: records the time it
takes to complete a problem set up to the millisecond.

Timer will pause and wait if you get a question wrong to give time to
check over your answer and will also pause once you exit the quesiton
page.
Timer can be hidden in case it induced stress.
Average times for vanilla problems will be logged and displayed in the
medals (achievement) page.
Best times for individual quesitons (vanilla) will also be logged and
displayed in the medals page.

Coming up:
Option to revert the time from a problem set if the average for that set
is much larger than your current (this can help if you some how leave
the computer and do something and forget to pause. You can delete the
instance of the logged time b/c it can "ruin" your average time). The
amount necessary to remove is subject to change.
Maybe some achievements for the special problem sets.
Option for the question to disappear after a set time to make the
problem more challenging.